+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-* Nothing
+* ```run-shellcheck.sh``` -> ```run-shelllint.sh```
 
 ### Removed